Experimenting with a faster but less precise skin-tone matching phase
This doesn't make a lot of difference for the small image included in
the app right now but it does make a big difference for larger images
that I've been experimenting with. The idea is that, if the SVM pass is
so cheap, it would be better if it was possible for the first pass to do
less work but potentially produce more possible face regions.